refactor: Логика отображения всех проектов вынесена с уровня страницы в фичу
