+++ src/features/show-projects/constants/index.ts
@@ -1,0 +1,1 @@
+export * from "./projects-tabs"
