+++ src/features/show-projects/types/types.ts
@@ -1,0 +1,1 @@
+export type TProjectsPageTabs = "activeProjects" | "waitListing" | "archive";
