+++ src/features/show-projects/ui/index.ts
@@ -2,3 +2,4 @@ export * from "./Project";
 export * from "./ActiveProjects";
 export * from "./ArchiveProjects";
 export * from "./WaitingProjects";
+export * from "./AllProjects";
