+++ src/shared/ui/Tabs/index.ts
@@ -1,1 +1,2 @@
 export * from "./Tabs";
+export * from "./Tab";
